resolve conflict
