+++ src/main.js
@@ -1,1 +1,2 @@
 import './js/script.js';
+import './js/animation.js';
